+++ cli2gui/__init__.py
@@ -1,1 +1,3 @@
+"""Entry point for the program.
+"""
 from .decorators import *